Trojan In Embedded PDF
This project focuses on creating a simple trojan using the metasploit payloads and exploits. When the victim opens this PDF file in their system the trojan will create a remote session on the attacker's machine to gain the complete control. This is related to ethical hacking use it on educational purpose only.